# api/openapi.yaml
openapi: 3.0.3
info:
  description: Fetch social media feeds
  title: Fetcher API
  version: 1.2.0
servers:
  - description: Fetcher Host
    url: https://jesse0michael-fetcher.herokuapp.com/
paths:
  /feed:
    get:
      description: Get feed.
      operationId: getFeed
      parameters:
        - description: twitterID
          in: query
          name: twitterID
          required: false
          schema:
            format: int64
            type: integer
        - description: instagramID
          in: query
          name: instagramID
          required: false
          schema:
            format: int64
            type: integer
        - description: bloggerID
          in: query
          name: bloggerID
          required: false
          schema:
            type: string
        - description: soundcloudID
          in: query
          name: soundcloudID
          required: false
          schema:
            type: string
        - description: swarmID
          in: query
          name: swarmID
          required: false
          schema:
            type: string
        - description: deviantartID
          in: query
          name: deviantartID
          required: false
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/feedItems"
          description: feed item array response
      summary: Get feed
  /proxy:
    get:
      summary: Proxy url
      description: Proxy url.
      operationId: proxy
      parameters:
        - description: url
          in: query
          name: url
          required: false
          schema:
            format: url
            type: string
      responses:
        "200":
          description: the proxied url content
components:
  schemas:
    feedItems:
      example:
        items:
          - id: CAfj92js93gjai
            ts: 1560834652
            source: instagram
          - id: hnoy78d34qdty3
            ts: 1560802637
            source: twitter
      properties:
        items:
          items:
            $ref: "#/components/schemas/feedItem"
          type: array
      required:
        - items
      type: object
    feedItem:
      example:
        id: hnoy78d34qdty3
        ts: 1560802637
        source: twitter
      properties:
        id:
          description: Unique identifier for a feed item
          type: string
        ts:
          description: Unix timestamp (seconds) for when the item was published
          format: int64
          type: integer
        source:
          description: The source platform the item is from
          enum:
            - twitter
            - instagram
            - blogger
            - soundcloud
            - swarm
            - deviantart
          type: string
        url:
          description: Permalink to the feed item on the platform
          type: string
        media:
          description: Array of media items (images, videos, etc...)
          items:
            $ref: "#/components/schemas/feedItem_media"
          type: array
        content:
          description: Text content for the item (may contain HTML)
          type: string
      required:
        - id
        - source
        - ts
      type: object
    feedItem_media:
      properties:
        url:
          description: The URL to the media content
          type: string
        poster:
          description: The URL to a poster image
          type: string
        kind:
          description: The kind of media
          enum:
            - image
            - video
            - audio
          type: string
      required:
        - kind
        - url